pyproject.toml
[project]
name = "tcli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "ipykernel>=7.2.0",
    "python-dotenv>=1.2.2",
    "telethon>=1.42.0",
    "toml>=0.10.2",
    "tqdm>=4.67.3",
    "typer>=0.24.1",
]

[dependency-groups]
dev = [
    "nuitka>=4.0.5",
    "zstandard>=0.25.0",
]
